+++ appdaemon/adapi.py
@@ -1824,6 +1824,18 @@ class ADAPI:
         self.logger.debug("list_services: %s", namespace)
         return self.AD.services.list_services(namespace)  # retrieve services
 
+    @overload # This overload provides the type hints for the Hass-specific version of this method
+    async def call_service(
+        self,
+        service: str,
+        namespace: str | None = None,
+        timeout: str | int | float | None = None,
+        callback: Callable | None = None,
+        hass_timeout: str | int | float | None = None,
+        suppress_log_messages: bool = False,
+        **data,
+    ) -> Any: ...
+
     @utils.sync_decorator
     async def call_service(
         self,
